test(client): split openai-compat helper tests into openai-compat.test.ts
The pure helpers in openai-compat.ts (parseOpenAICompatEnvelope,
collectSystemFromMessages, chatHistoryFromMessages,
buildOpenAICompatSystemPrompt, callerToolDispatchActive,
tryParseToolCallsEnvelope, formatChatHistory, dumpOpenAICompatTaskWire)
have historically been covered from claude.test.ts — a holdover from
when the helpers lived alongside the claude backend code. With #302
closing out and the helpers now having no special relationship to
claude, move every helper test out into a dedicated openai-compat.test.ts
that lives next to the module it covers.

claude.test.ts keeps the claude-specific tests:

  - envelope.model → `--model` spawn argv forwarding (#302)
  - buildOpenAICompatNativeSystemPrompt (the claude-side native-MCP
    system prompt, separate from buildOpenAICompatSystemPrompt)
  - argv composition, native dispatch end-to-end, capabilities probe,
    session reuse, traceability artifacts

No behaviour change — the migrated tests are identical to the originals
(only the trailing block of more-comprehensive tryParseToolCallsEnvelope
cases got expanded slightly so the new file pins the full surface in
one place).

Test counts:
- claude.test.ts: 128 → 98 (helper tests moved out)
- openai-compat.test.ts: 30 (new)
- Overall: 622 client tests pass.
